refact: polish detail
